Refactoring API method so query can be made through argument
