+++ src/app/mock-pub.ts
@@ -2,9 +2,9 @@ import { Restaurant } from './restaurant';
 
 export const PUBS: Restaurant[] = [
     {
-        id: 4,
+        id: '4',
         name: 'The Horniman at Hays',
-        rating: 4.0,
+        rating: '4.0',
         category: 'Pub',
         address: 'Hay’s Galleria, Unit 26, London SE1 2HD',
         url: 'nicholsonspubs.co.uk',
