+++ src/app/mock-restaurants.ts
@@ -2,17 +2,17 @@ import { Restaurant } from './restaurant';
 
 export const RESTAURANTS: Restaurant[] = [
     {
-        id: 6,
+        id: '6',
         name: 'Gaucho',
-        rating: 4.3,
+        rating: '4.3',
         category: 'Argentinian restaurant',
         address: '2 More London Riverside, London SE1 2AP',
         url: 'gauchorestaurants.com',
     },
     {
-        id: 7,
+        id: '7',
         name: 'La Dame de Pic London',
-        rating: 4.7,
+        rating: '4.7',
         category: 'French restaurant',
         address: '10 Trinity Square, London EC3N 4AJ',
         url: 'ladamedepiclondon.co.uk',
